pyproject.toml
[project]
name = "pivot-navigation"
version = "0.1.0"
description = "PIVOT: Prompting with Iterative Visual Optimization for Drone Navigation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "airsim>=1.8.1",
    "backports-ssl-match-hostname>=3.7.0.1",
    "djitellopy>=2.5.0",
    "google-genai>=1.35.0",
    "matplotlib>=3.10.6",
    "mss>=10.1.0",
    "numpy>=2.3.3",
    "openai>=1.107.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]

[project.scripts]
pivot = "pivot.main:main"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.uv.extra-build-dependencies]
airsim = ["numpy", "msgpack-rpc-python", "backports.ssl_match_hostname"]
